generate rpn proposal file debug

+++ relation_rcnn/function/train_rpn.py
@@ -121,7 +121,7 @@ def train_rpn(cfg, dataset, image_set, root_path, dataset_path,
     # decide learning rate
     base_lr = lr
     lr_factor = cfg.TRAIN.lr_factor
-    lr_epoch = [int(epoch) for epoch in lr_step.split(',')]
+    lr_epoch = [float(epoch) for epoch in lr_step.split(',')]
     lr_epoch_diff = [epoch - begin_epoch for epoch in lr_epoch if epoch > begin_epoch]
     lr = base_lr * (lr_factor ** (len(lr_epoch) - len(lr_epoch_diff)))
     lr_iters = [int(epoch * len(roidb) / batch_size) for epoch in lr_epoch_diff]
